Move do_prefetch() from easyaccess to do_utils file. Move _complete_tables(), _complete_colnames(), complete_prefetch(), and complete_shell() fuctions to dbdo_utils file
